Re-implement build_path in a more efficient way

+++ src/string.h
@@ -10,7 +10,7 @@
 #include <string.h>
 
 char *whiteout_tag(const char *fname);
-int build_path(char *dest, int max_len, const char *callfunc, int line, ...);
+int build_path(char *dest, size_t max_len, const char *callfunc, int line, ...);
 char *u_dirname(const char *path);
 unsigned int string_hash(void *s);
